refact: extrair handlers de merchant trade para module/trade/merchant-trade-handlers.mjs
